requirements.txt
# Ingestion
requests==2.31.0
boto3==1.34.0

# Orchestration
apache-airflow==2.8.1

# Databricks
databricks-sdk==0.20.0
databricks-sql-connector==3.0.0

# Dashboard
streamlit==1.32.0
plotly==5.20.0
pandas==2.2.0

# Utilities
python-dotenv==1.0.1
